collect inscription relevant structs into module
